+++ src/main/java/org/prography/spring/dto/responseDto/RoomListResponse.java
@@ -11,11 +11,11 @@ import java.util.List;
 @AllArgsConstructor(access = lombok.AccessLevel.PRIVATE)
 public class RoomListResponse {
 
-    private Integer totalElements;
-    private Integer totalPages;
+    private Long totalElements;
+    private Long totalPages;
     private List<RoomResponse> rooms;
 
-    public static RoomListResponse of(Integer totalElements, Integer totalPages, List<RoomResponse> rooms) {
+    public static RoomListResponse of(Long totalElements, Long totalPages, List<RoomResponse> rooms) {
         return RoomListResponse.builder()
                 .totalElements(totalElements)
                 .totalPages(totalPages)
